backend/02_ai_model_tutorial/backend/pyproject.toml
[project]
name = "python-pet-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.5",
    "google-genai>=1.53.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
